% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{accuracy}
\alias{accuracy}
\title{Compute accuracy metric}
\usage{
accuracy(predictions, true_labels)
}
\arguments{
\item{predictions}{Vector of predicted labels.}

\item{true_labels}{Vector of true labels.}
}
\value{
Accuracy value.
}
\description{
Compute accuracy metric
}
