% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDistance.R
\name{CalculateDistance}
\alias{CalculateDistance}
\title{Calculates euclidean distance between 2D data points}
\usage{
CalculateDistance(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
distance: euclidean distance between points, real
}
\description{
Calculates euclidean distance between 2D data points
}
\examples{
CalculateDistance(c(1,2), c(3,4))
}
