% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int_centroid_matrix.R
\name{centroid_matrix}
\alias{centroid_matrix}
\title{centroid_matrix}
\usage{
centroid_matrix(x, m)
}
\arguments{
\item{x}{\code{x} A vector of integer or double with elements \code{n >= 1}.}

\item{m}{\code{m} An integer declaring the number of unlabelled training examples.}
}
\description{
\code{centroid_matrix} rowNorms
}
\examples{
# X is a matrix of unlabelled training examples

# X <- matrix(rep( 0, 10), ncol = 2)

# y is a vector of initial centroids

centroids <- c(1,10)

# X - centroid_matrix(x = centroids, m = nrow(X))
}

