% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/centroid_mean.R
\name{centroid_mean}
\alias{centroid_mean}
\title{Calculate centroid mean}
\usage{
centroid_mean(X, group)
}
\arguments{
\item{X}{\code{X} Matrix of \code{ncol >=2}.}

\item{group}{\code{group} Vector of groups for which centroid means will be calculated.}
}
\description{
\code{centroid_mean} calculate centroid mean
}
\examples{
#centroid_mean(X, idx)
}

