% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean Distance}
\usage{
euclidean_distance(x1, x2)
}
\arguments{
\item{x1}{numeric variable}

\item{x2}{numeric variable}
}
\value{
Euclidean distance value
}
\description{
This function will help you perform Euclidean Distance.
}
