% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{f1_score}
\alias{f1_score}
\title{Compute F1 score metric}
\usage{
f1_score(predictions, true_labels, positive_class)
}
\arguments{
\item{predictions}{Vector of predicted labels.}

\item{true_labels}{Vector of true labels.}

\item{positive_class}{Label of the positive class.}
}
\value{
F1 score value.
}
\description{
Compute F1 score metric
}
