% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Calculate nearest centroid}
\usage{
find_centroid(X, centroids = NULL, k = NULL)
}
\description{
\code{find_centroid} calculate nearest centroid
}
\examples{
find_centroid(X)
}

