% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_group.R
\name{find_group}
\alias{find_group}
\title{Calculate group}
\usage{
find_group(X, centroids = NULL, k = 3)
}
\arguments{
\item{X}{Matrix of \code{ncol >=2}.}

\item{centroids}{Vector of initial centroid locations. Defaults to \code{NULL}.}

\item{k}{Number of randomly initiliased centroids to create. Deafults to 3, and is required in \code{centroids = NULL}.}
}
\description{
\code{find_group} calculate nearest centroid
}
\examples{
X <- mtcars[, c("disp", "mpg")]
idx <- find_group(X, k = 2)
}

