% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyKNN.R
\name{FindCategory}
\alias{FindCategory}
\title{Apply KNN}
\usage{
FindCategory(training.data, test.point, k)
}
\arguments{
\item{training.data}{training data table where columns 1 and 2 are 
Euclidean coordinates (numeric), and column 3 is the category}

\item{test.data}{2D test point, as Euclidean coordinates (numeric)}
}
\value{
knn.table :table of test data and KNN predictions
}
\description{
This function predicts categories on a 2D test data set using training data.
}
\examples{
iris.train <- iris[1:120,c(1,4,5)]
iris.test <- iris[121:150,c(1,4)]
ApplyKNN(iris.train, iris.test, 5)
}
\keyword{KNN}
