% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{fit.internal}
\alias{fit.internal}
\title{Internal Fitting Function}
\usage{
fit.internal(
  phenotype,
  genotype,
  input.kernel,
  MINQUE.type,
  KNN.type,
  lambda,
  weight,
  fixed.effects,
  constrain
)
}
\description{
Internal function that performs the actual KNN fitting without CV.
This is used within the CV loop to avoid infinite recursion.
}
