% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{GKNN.predict}
\alias{GKNN.predict}
\title{GKNN Prediction}
\usage{
GKNN.predict(
  phenotype,
  genotype,
  theta,
  random.phenotype,
  covariance.inverse,
  input.kernel = NULL,
  family = "binomial",
  KNN.type = "KNN",
  beta = NULL,
  fixed.effects.train = NULL,
  fixed.effects.test = NULL,
  theta.nb = NULL,
  pi.zinb = NULL
)
}
\arguments{
\item{phenotype}{Response variable (phenotype) of training subjects.}

\item{genotype}{List of N x p genotype matrices (combined training and testing subjects).}

\item{theta}{Estimated variance components.}

\item{random.phenotype}{Estimated random effects.}

\item{covariance.inverse}{Estimated inverse covariance matrix of random effects.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{family}{The exponential family ('binomial', 'poisson', 'negbinomial', 'zinb') (default 'binomial').}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM') (default 'KNN').}

\item{beta}{Estimated fixed effects coefficients.}

\item{fixed.effects.train}{Fixed effects design matrix for training subjects.}

\item{fixed.effects.test}{Fixed effects design matrix for testing subjects.}

\item{theta.nb}{Overdispersion parameter for negative binomial (used if family is 'negbinomial' or 'zinb').}

\item{pi.zinb}{Zero-inflation probability for ZINB (used if family is 'zinb').}
}
\value{
Predicted phenotypes for testing subjects.
}
\description{
Predicts phenotypes for new subjects using the GKNN model.
}
