% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN.ind.test}
\alias{KNN.ind.test}
\title{KNN Mixture Chi-Square Individual Test}
\usage{
KNN.ind.test(
  phenotype,
  genotype,
  test.index,
  input.kernel = NULL,
  MINQUE.type = "MINQUE0",
  KNN.type = "KNN",
  lambda = 0,
  weight = NULL,
  fixed.effects = NULL,
  constrain = FALSE,
  variance.component.list = NULL
)
}
\arguments{
\item{phenotype}{Numeric vector of response variable (phenotype) of length N.}

\item{genotype}{List of N x p genotype matrices.}

\item{test.index}{Index of the variance component to test.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{MINQUE.type}{Type of MINQUE estimator ('MINQUE0' or 'MINQUE1') (default is 'MINQUE0').}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM').}

\item{lambda}{Regularization parameter (default is 0).}

\item{weight}{Optional weight vector for variance components.}

\item{fixed.effects}{Optional fixed effects design matrix.}

\item{constrain}{Logical indicating whether to constrain variance components to be non-negative (default is FALSE).}

\item{variance.component.list}{Optional precomputed list of variance components.}
}
\value{
A list containing the p-value of the test.
}
\description{
Performs an individual test for a variance component in a KNN model using a mixture of chi-square distributions.
}
