% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN.predict}
\alias{KNN.predict}
\title{KNN Prediction}
\usage{
KNN.predict(
  phenotype,
  genotype,
  theta,
  input.kernel = NULL,
  KNN.type = "KNN",
  beta = NULL,
  fixed.effects.train = NULL,
  fixed.effects.test = NULL,
  variance.component.list = NULL
)
}
\arguments{
\item{phenotype}{Response variable (phenotype) of training subjects.}

\item{genotype}{List of N x p genotype matrices (combined training and testing subjects).}

\item{theta}{Estimated variance components.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM') (default is 'KNN').}

\item{beta}{Estimated fixed effects coefficients.}

\item{fixed.effects.train}{Fixed effects design matrix for training subjects.}

\item{fixed.effects.test}{Fixed effects design matrix for testing subjects.}

\item{variance.component.list}{Optional precomputed list of variance components.}
}
\value{
Predicted phenotypes for testing subjects.
}
\description{
Predicts phenotypes for new subjects using the KNN model.
}
