% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{knn}
\alias{knn}
\title{KNN model function}
\usage{
knn(
  train_data,
  test_data,
  target_train,
  k = 5,
  distance = "euclidean",
  minkowski_p = 2,
  task = "classification"
)
}
\arguments{
\item{train_data}{The training data.}

\item{test_data}{The test data.}

\item{target_train}{The target variable in the training data.}

\item{k}{The number of neighbors to consider (default is 5).}

\item{distance}{The distance metric to use (default is "euclidean").}

\item{minkowski_p}{The power parameter for Minkowski distance (default is 2).}

\item{task}{The task type, either "classification" or "regression" (default is "classification").}
}
\value{
A vector of predictions for the test data.
}
\description{
This function implements the K-nearest neighbors algorithm for classification.
}
