% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN2LMM}
\alias{KNN2LMM}
\title{KNN to LMM Conversion}
\usage{
KNN2LMM(genotype, input.kernel = NULL, KNN.type = "KNN")
}
\arguments{
\item{genotype}{List of N x p genotype matrices.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM').}
}
\value{
List of variance component matrices.
}
\description{
Converts kernel neural networks into linear mixed models by constructing variance component matrices.
}
