% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{knn_cv}
\alias{knn_cv}
\title{use cross-validation to tune hyperparameter(k_folds)}
\usage{
knn_cv(data, target, k_folds, k_neighbors, distance_function)
}
\arguments{
\item{data}{dataset use here}

\item{target}{the dependent variable to predict}

\item{k_folds}{number of folds in cross-validation}

\item{k_neighbors}{number of neighbours in a group}

\item{distance_function}{such as Euclidean distance,Manhattan distance and Minkowski distance}
}
\value{
accuracy
}
\description{
use cross-validation to tune hyperparameter(k_folds)
}
\examples{
library(tidyr)
library(palmerpenguins)
library(dplyr)
data("penguins")
k_folds <- 5
k_neighbors <- 5
euclidean_distance = function(a, b){if(length(a) == length(b)){sqrt(sum((a-b)^2))  } else{stop('Vectors must be of the same length')}}
knn_cv(penguins, "species", k_folds, k_neighbors, euclidean_distance)
}
