% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{knn_imputation}
\alias{knn_imputation}
\title{KNN imputation function}
\usage{
knn_imputation(data, k = 3, distance = "euclidean", minkowski_p = 2)
}
\arguments{
\item{data}{The input data.}

\item{k}{The number of neighbors to consider for imputation (default k=3).}

\item{distance}{The distance method (defautl is 'euclidean')}

\item{minkowski_p}{The power parameter for Minkowski distance (default is 2).}
}
\value{
The imputed dataset.
}
\description{
This function performs KNN imputation for missing values in a dataset.
}
