% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{knn_model}
\alias{knn_model}
\title{KNN model wrapper}
\usage{
knn_model(train_data, target_train, k = 5, distance = "euclidean")
}
\arguments{
\item{train_data}{The training data.}

\item{target_train}{The target variable in the training data.}

\item{k}{The number of neighbors to consider (default is 5).}

\item{distance}{The distance metric to use (default is "euclidean").}
}
\value{
An object representing the KNN model.
}
\description{
This function wraps the KNN model for convenient use.
}
