% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_functions.R
\name{knn_predict}
\alias{knn_predict}
\title{KNN Predict Function}
\usage{
knn_predict(formula, data, newdata, k = 1, distance_method = "euclidean")
}
\arguments{
\item{formula}{A formula of the form class ~ predictors}

\item{data}{Training data}

\item{newdata}{Test data}

\item{k}{Number of neighbors}

\item{distance_method}{Distance method}
}
\value{
Factor of predictions
}
\description{
A more user-friendly interface for KNN predictions
}
\examples{
# Formula interface with iris dataset
data(iris)
set.seed(123)
train_idx <- sample(1:nrow(iris), 100)
train_data <- iris[train_idx, ]
test_data <- iris[-train_idx, ]
pred <- knn_predict(Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
                    train_data, test_data, k = 3)
table(pred, test_data$Species)

# Using only specific predictors
pred_subset <- knn_predict(Species ~ Sepal.Length + Petal.Length,
                          train_data, test_data, k = 3)
}
