% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knn_Reg.R
\name{knn_reg}
\alias{knn_reg}
\title{K-nearest Neighbors Algorithm for Regression}
\usage{
knn_reg(train_data, test_data, k = 3)
}
\arguments{
\item{train_data}{training set of your data}

\item{test_data}{test set of your data}

\item{k}{number of neighbors(default is 3)}
}
\value{
Predicted response for regression data
}
\description{
This function will help you perform K-Nearest Neighbors Algorithm on
regression data. The default distance that was used is euclidean distance.
And the default parameter k is 3. It will returned the stored prediction of the response
only if you have your data make sure the response is in the last
column and before call out the knn function you need to separate your data into train and test set.
}
\examples{
# load iris dataset
data(faithful)
head(faithful)
# Split data into training and test sets
set.seed(123)
train_rows <- sample(1:nrow(faithful), 0.7 * nrow(faithful))
train_data <- faithful[train_rows, ]
test_data <- faithful[-train_rows, ]
predicted_values <- knn_reg(train_data, test_data, k = 3)
predicted_values
actual_values <- test_data[, ncol(test_data)]
# Compute RMSE
rmse <- sqrt(mean((predicted_values - actual_values) ^ 2))
cat("RMSE:", rmse, "\n")
}
