\name{knn_singleton}
\alias{knn_singleton}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	knn_singleton
}
\description{
	see "knn"
}
\usage{
knn_singleton(m1 = 0, m2 = 0, e = 0, dgp1 = "1", dgp2 = "1", n = 20, loops = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m1}{
see "knn"
}
  \item{m2}{
see "knn"
}
  \item{e}{
see "knn"
}
  \item{dgp1}{
see "knn"
}
  \item{dgp2}{
see "knn"
}
  \item{n}{
see "knn"
}
  \item{loops}{
see "knn"
}
}
\details{
see "knn"
}
\value{
see "knn"
}
\references{
see "knn"
}
\author{
Ernest Liu, Daisy Sun, Aaron Swoboda.
}
\note{
see "knn"
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
knn_singleton(0,0,0)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
