% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{mode}
\alias{mode}
\title{Mode function}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Should missing values be removed? (default is FALSE).}
}
\value{
The mode of the vector.
}
\description{
This function calculates the mode of a vector.
}
