% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{MVKNN.ind.test}
\alias{MVKNN.ind.test}
\title{MVKNN Mixture Chi-Square Individual Test}
\usage{
MVKNN.ind.test(
  phenotype,
  genotype,
  test.index,
  input.kernel = NULL,
  MINQUE.type = "MINQUE0",
  KNN.type = "KNN",
  phenotype.type = "Heter",
  lambda = 0,
  weight = NULL,
  fixed.effects = NULL,
  constrain = TRUE,
  thresh = 1e-05,
  max.iter = 100
)
}
\arguments{
\item{phenotype}{Matrix of phenotypes (N x d), where N is the number of samples and d is the number of traits.}

\item{genotype}{List of N x p genotype matrices.}

\item{test.index}{Index of the variance component(s) to test.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{MINQUE.type}{Type of MINQUE estimator ('MINQUE0' or 'MINQUE1') (default 'MINQUE0').}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM') (default 'KNN').}

\item{phenotype.type}{Phenotype covariance structure ('Homo' or 'Heter') (default 'Heter').}

\item{lambda}{Regularization parameter (default 0).}

\item{weight}{Optional weight vector for variance components.}

\item{fixed.effects}{Optional fixed effects design matrix.}

\item{constrain}{Logical indicating whether to constrain variance components to be non-negative (default FALSE).}

\item{thresh}{Convergence threshold for iterative methods (default 1e-5).}

\item{max.iter}{Maximum number of iterations (default 100).}
}
\value{
A list containing the p-value of the test.
}
\description{
Performs an individual test for variance components in a multivariate KNN model using a mixture of chi-square distributions.
}
