% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{MVKNN.predict}
\alias{MVKNN.predict}
\title{MVKNN Prediction}
\usage{
MVKNN.predict(
  phenotype,
  genotype,
  theta,
  phenotype.covariance,
  input.kernel = NULL,
  KNN.type = "KNN",
  beta = NULL,
  fixed.effects.train = NULL,
  fixed.effects.test = NULL
)
}
\arguments{
\item{phenotype}{Matrix of training phenotypes (N.train x d), where N.train is the number of training samples and d is the number of traits.}

\item{genotype}{List of N x p genotype matrices (combined training and testing subjects).}

\item{theta}{Estimated variance components.}

\item{phenotype.covariance}{Estimated covariance matrix of phenotypes (d x d).}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM') (default 'KNN').}

\item{beta}{Estimated fixed effects coefficients (optional).}

\item{fixed.effects.train}{Fixed effects design matrix for training subjects (optional).}

\item{fixed.effects.test}{Fixed effects design matrix for testing subjects (optional).}
}
\value{
Predicted phenotypes for testing subjects as a matrix (N.test x d).
}
\description{
Predicts multivariate phenotypes for new subjects using the MVKNN model.
}
