% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\docType{data}
\name{num_cores}
\alias{num_cores}
\title{Detect the number of available CPU cores}
\format{
An object of class \code{integer} of length 1.
}
\usage{
num_cores
}
\value{
An integer representing the number of CPU cores.
}
\description{
Use the detectCores function to determine the number of CPU cores available.
}
\keyword{datasets}
