% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{perform.cv}
\alias{perform.cv}
\title{Internal Cross-Validation Function}
\usage{
perform.cv(
  phenotype,
  genotype,
  input.kernel,
  MINQUE.type,
  KNN.type,
  lambda.list,
  weight,
  fixed.effects,
  constrain,
  criteria
)
}
\arguments{
\item{phenotype}{Numeric vector of response variable (phenotype) of length N.}

\item{genotype}{List of N x p genotype matrices.}

\item{input.kernel}{List specifying the kernel function(s) for each genotype matrix.}

\item{MINQUE.type}{Type of MINQUE estimator ('MINQUE0' or 'MINQUE1').}

\item{KNN.type}{Type of KNN model ('KNN', 'LMM').}

\item{lambda.list}{List of lambda values to try.}

\item{weight}{Optional weight vector for variance components.}

\item{fixed.effects}{Optional fixed effects design matrix.}

\item{constrain}{Logical indicating whether to constrain variance components to be non-negative.}

\item{criteria}{Criterion for selecting optimal lambda ('MSE', 'COR').}

\item{variance.component.list}{Optional precomputed list of variance components.}
}
\value{
The optimal lambda value.
}
\description{
Internal function to perform cross-validation for lambda selection.
This function is called internally by KNN when lambda is not specified.
}
