% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{perform.gknn.cv}
\alias{perform.gknn.cv}
\title{GKNN Cross-Validation Function}
\usage{
perform.gknn.cv(
  phenotype,
  genotype,
  input.kernel,
  family,
  MINQUE.type,
  KNN.type,
  lambda.list,
  weight,
  beta,
  fixed.effects,
  constrain,
  criteria,
  thresh,
  max.iter,
  alpha,
  lower.limits,
  numerical,
  theta.nb = NULL,
  pi.zinb = NULL
)
}
\description{
Performs cross-validation for lambda selection in GKNN models.
}
