% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{perform.mvknn.cv}
\alias{perform.mvknn.cv}
\title{MVKNN Cross-Validation Function}
\usage{
perform.mvknn.cv(
  phenotype,
  genotype,
  input.kernel,
  MINQUE.type,
  KNN.type,
  phenotype.type,
  lambda.list,
  weight,
  fixed.effects,
  constrain,
  criteria,
  thresh,
  max.iter
)
}
\description{
Performs cross-validation for lambda selection in MVKNN models.
}
