% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_knn.R
\name{plot_knn}
\alias{plot_knn}
\title{Plot k-nearest neighbours}
\usage{
plot_knn(X, centroids, idx = NULL)
}
\description{
\code{plot_knn} plot k nearest neighbours
}
\examples{
plot(X, centroids, idx)
}

