% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_knn_path.R
\name{plot_knn_path}
\alias{plot_knn_path}
\title{Plot centroids paths}
\usage{
plot_knn_path(X, centroids_all)
}
\arguments{
\item{X}{\code{X} Matrix of \code{ncol >=2}.}

\item{centroids_all}{\code{centroids_all} An input dataframe comprising three variable: \code{x}, \code{y} are coordinates of the locations of centroids, \code{i} is number of the centroid.}
}
\description{
\code{plot_knn_path} Plot centroids paths
}
\examples{
#plot_knn_path( X, centroids_all)
}

