% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{precision}
\alias{precision}
\title{Compute precision metric}
\usage{
precision(predictions, true_labels, positive_class)
}
\arguments{
\item{predictions}{Vector of predicted labels.}

\item{true_labels}{Vector of true labels.}

\item{positive_class}{Label of the positive class.}
}
\value{
Precision value.
}
\description{
Compute precision metric
}
