% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{predict.knn_model}
\alias{predict.knn_model}
\title{Predict method for KNN model}
\usage{
\method{predict}{knn_model}(model, new_data)
}
\arguments{
\item{model}{The KNN model object.}

\item{new_data}{The new data for prediction.}
}
\value{
A vector of predictions for the new data.
}
\description{
This function predicts the target variable using the KNN model.
}
