% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{r_cv}
\alias{r_cv}
\title{Repeated Cross-Validation Function}
\usage{
r_cv(
  data,
  target_col,
  k_values = c(3, 5),
  test_size = 0.2,
  distance_metric = "euclidean",
  num_folds = 5,
  Scale = TRUE,
  minkowski_p = 2,
  task = "classification",
  positive_class = NULL
)
}
\arguments{
\item{data}{The input data.}

\item{target_col}{The target column in the dataset.}

\item{k_values}{A vector of positive integers representing different values of k.}

\item{test_size}{The proportion of the data to use for testing (default is 0.2).}

\item{distance_metric}{The distance metric to use (default is "euclidean").}

\item{num_folds}{The number of folds for cross-validation (default is 5).}

\item{Scale}{Should numeric variables be scaled? (default is TRUE).}

\item{minkowski_p}{The power parameter for Minkowski distance (default is 2).}

\item{task}{The task type, either "classification" or "regression" (default is "classification").}

\item{positive_class}{The positive class from the target col used to caluculate different metrics(default is NULL).}
}
\value{
A list containing mean accuracies and predictions for each k value.
}
\description{
This function performs repeated cross-validation for KNN classification and regression.
}
