% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_knn.R
\name{recall}
\alias{recall}
\title{Compute recall metric}
\usage{
recall(predictions, true_labels, positive_class)
}
\arguments{
\item{predictions}{Vector of predicted labels.}

\item{true_labels}{Vector of true labels.}

\item{positive_class}{Label of the positive class.}
}
\value{
Recall value.
}
\description{
Compute recall metric
}
