% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int_rowNorms.R
\name{rowNorms}
\alias{rowNorms}
\title{rowNorms}
\usage{
rowNorms(x)
}
\arguments{
\item{\code{x}}{A matrix object of \code{nrows >= 1} and \code{ncol >= 2}.}
}
\description{
\code{rowNorms} rowNorms
}
\examples{
a <- matrix(1:20,ncol=2)

rowNorms(a)
}

