% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Calculate nearest centroid}
\usage{
find_centroid(X, centroids = NULL, k = NULL)
}
\arguments{
\item{X}{\code{X} Matrix of \code{ncol >=2}.}

\item{centroids}{\code{centroids} Vector of initial centroid locations. Defaults to \code{NULL}.}

\item{k}{\code{k} Number of randomly initiliased centroids to create. Deafults to \code{NULL}, but is required in \code{centroids = NULL}.}
}
\description{
\code{find_centroid} calculate nearest centroid
}
\examples{
#find_centroid(X)
}

