% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{knn}
\alias{knn}
\title{Imitate knn}
\usage{
knn(train_data, train_labels, test_data, k, distance_function)
}
\arguments{
\item{train_data}{split the original data randomly into 2 datasets, one is training dataset, train data is its features}

\item{train_labels}{split the original data randomly into 2 datasets, one is training dataset, train_labels is its target}

\item{test_data}{split the original data randomly into 2 datasets, another one is testing dataset, test_data is its features}

\item{k}{number of neighbours in a group}

\item{distance_function}{such as Euclidean distance,Manhattan distance and Minkowski distance}
}
\value{
prediction of the target variable
}
\description{
Imitate knn
}
\examples{
library(tidyr)
library(palmerpenguins)
library(tidyverse)
library(dplyr)
data("penguins")
penguins <- na.omit(penguins)
penguins <- penguins \%>\% select(-year)
features <- penguins[, c("bill_length_mm", "bill_depth_mm", "flipper_length_mm", "body_mass_g")]
target <- penguins$species
set.seed(123)
indices <- sample(1:nrow(features), size = 0.7 * nrow(features))
train_x <- features[indices, ]
train_y <- target[indices]
test_x <- features[-indices, ]
test_y <- target[-indices]
k <- 5
euclidean_distance = function(a, b){if(length(a) == length(b)){sqrt(sum((a-b)^2))  } else{stop('Vectors must be of the same length')}}
predictions <- knn(train_x, train_y, test_x, k, euclidean_distance)
}
