% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{Calculate centroid mean}
\usage{
knn(X, centroids = NULL, k = NULL, imax = NULL, ...)
}
\arguments{
\item{X}{\code{X} Matrix of \code{ncol >=2}.}

\item{group}{\code{group} Vector of groups for which centroid means will be calculated.}
}
\description{
\code{centroid_mean} calculate centroid mean
}
\examples{
# centroid_mean(X, idx)
}

