% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_cpp_wrapper.R
\name{knn_cpp_wrapper}
\alias{knn_cpp_wrapper}
\title{K-Nearest Neighbors Classification using Rcpp
This function implements the K-Nearest Neighbors algorithm for classification
using Rcpp for improved performance.}
\usage{
knn_cpp_wrapper(train, test, cl, k = 1, distance_method = "euclidean")
}
\description{
K-Nearest Neighbors Classification using Rcpp
This function implements the K-Nearest Neighbors algorithm for classification
using Rcpp for improved performance.
}
